// ui/src/api/generated/model/serverDiskInfo.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Whirlpool PDI Tool API
 * Whirlpool PDI Tool API APIs developed by Scopt Analytics.
 * OpenAPI spec version: 1.7.1
 */

export interface ServerDiskInfo {
  mountpoint: string;
  device: string;
  fstype: string;
  total_bytes: number;
  used_bytes: number;
  free_bytes: number;
  percent: number;
}